requirements.txt
#
# This file contains a list of external dependencies
# needed to run stoqdrivers, they are processed by pip/easy_install
#

kiwi-gtk >= 1.9.41
zope.interface >= 3.0
pyserial >= 2.2
pyusb
qrcode >= 4.0.1
